utils_test: fix isqrt_test calculation range
Multiplying the uint16_t x by itself seems to default to be calculated in
int32_t range, while it obviously needs uint32_t. This causes sporadic
sanitizer barfs:

    Testing integer square-root
    utils_test.c:445:18: runtime error: signed integer overflow: 60369 * 60369 cannot be represented in type 'int'

The final result is still correct, because it is in fact interpreted as uint32_t.

Cast to uint32_t to make sure the sanitizer doesn't complain.

Related: OS#3407

+++ tests/utils/utils_test.c
@@ -439,7 +439,7 @@ static void isqrt_test(void)
 			x = r * (UINT16_MAX/RAND_MAX);
 		else
 			x = r;
-		uint32_t sq = x*x;
+		uint32_t sq = (uint32_t)x*x;
 		uint32_t y = osmo_isqrt32(sq);
 		if (y != x)
 			printf("ERROR: x=%u, sq=%u, osmo_isqrt(%u) = %u\n", x, sq, sq, y);
